+++ data.go
@@ -15,18 +15,24 @@ import (
 )
 
 type DataPageData struct {
-	Version        string
-	Company        string
-	Alarms         string
-	MenuOverview   string
-	MenuWorkplaces string
-	MenuCharts     string
-	MenuStatistics string
-	MenuData       string
-	MenuSettings   string
-	SelectionMenu  []string
-	Workplaces     []string
-	Compacted      string
+	Version               string
+	Company               string
+	Alarms                string
+	MenuOverview          string
+	MenuWorkplaces        string
+	MenuCharts            string
+	MenuStatistics        string
+	MenuData              string
+	MenuSettings          string
+	SelectionMenu         []Selection
+	Workplaces            []string
+	Compacted             string
+	DataFilterPlaceholder string
+	DateLocale            string
+}
+type Selection struct {
+	SelectionName string
+	Selection     string
 }
 
 type DataPageInput struct {
@@ -43,6 +49,8 @@ type DataPageOutput struct {
 
 type OrderRecord struct {
 	WorkplaceName     string
+	OrderStart        string
+	OrderEnd          string
 	WorkplaceModeName string
 	WorkshiftName     string
 	UserName          string
@@ -56,8 +64,11 @@ type OrderRecord struct {
 }
 
 type TableData struct {
-	TableHeader []HeaderCell
-	TableRows   []TableRow
+	DataTableSearchTitle    string
+	DataTableInfoTitle      string
+	DataTableRowsCountTitle string
+	TableHeader             []HeaderCell
+	TableRows               []TableRow
 }
 type TableRow struct {
 	TableCell []TableCell
@@ -77,6 +88,7 @@ func data(writer http.ResponseWriter, request *http.Request, _ httprouter.Params
 	email, _, _ := request.BasicAuth()
 	var data DataPageData
 	data.Version = version
+	data.DateLocale = getLocaleForPickers(email)
 	data.Company = cachedCompanyName
 	data.MenuOverview = getLocale(email, "menu-overview")
 	data.MenuWorkplaces = getLocale(email, "menu-workplaces")
@@ -84,19 +96,52 @@ func data(writer http.ResponseWriter, request *http.Request, _ httprouter.Params
 	data.MenuStatistics = getLocale(email, "menu-statistics")
 	data.MenuData = getLocale(email, "menu-data")
 	data.MenuSettings = getLocale(email, "menu-settings")
+	data.DataFilterPlaceholder = getLocale(email, "data-table-search-title")
 	data.Compacted = cachedUserSettings[email].menuState
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "alarms"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "breakdowns"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "downtimes"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "faults"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "orders"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "packages"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "parts"))
-	data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "states"))
+	fmt.Println(cachedUserSettings[email].dataSelection)
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "alarms"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "alarms"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "breakdowns"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "breakdowns"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "downtimes"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "downtimes"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "faults"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "faults"),
+	})
+
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "orders"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "orders"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "packages"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "packages"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "parts"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "parts"),
+	})
+	data.SelectionMenu = append(data.SelectionMenu, Selection{
+		SelectionName: getLocale(email, "states"),
+		Selection:     getSelected(cachedUserSettings[email].dataSelection, "states"),
+	})
 	if cachedUsersByEmail[email].UserTypeID == 2 {
 		logInfo("MAIN", "Adding data menu for administrator")
-		data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "users"))
-		data.SelectionMenu = append(data.SelectionMenu, getLocale(email, "system-statistics"))
+		data.SelectionMenu = append(data.SelectionMenu, Selection{
+			SelectionName: getLocale(email, "users"),
+			Selection:     getSelected(cachedUserSettings[email].dataSelection, "users"),
+		})
+		data.SelectionMenu = append(data.SelectionMenu, Selection{
+			SelectionName: getLocale(email, "system-statistics"),
+			Selection:     getSelected(cachedUserSettings[email].dataSelection, "system-statistics"),
+		})
 	}
 
 	db, err := gorm.Open(postgres.Open(config), &gorm.Config{})
@@ -117,6 +162,68 @@ func data(writer http.ResponseWriter, request *http.Request, _ httprouter.Params
 	logInfo("MAIN", "Home page sent")
 }
 
+func getSelected(selection string, menu string) string {
+	fmt.Println(selection + " - " + menu)
+	if selection == menu {
+		fmt.Println("match")
+		return "selected"
+	}
+	return ""
+}
+
+func getLocaleForPickers(email string) string {
+	var pickerLocale string
+	user, _ := cachedUsersByEmail[email]
+	switch user.Locale {
+	case "CsCZ":
+		{
+			pickerLocale = "cs-CZ"
+		}
+	case "DeDE":
+		{
+			pickerLocale = "de-DE"
+		}
+	case "EnUS":
+		{
+			pickerLocale = "en-US"
+		}
+	case "EsES":
+		{
+			pickerLocale = "es-MX"
+		}
+	case "FrFR":
+		{
+			pickerLocale = "fr-FR"
+		}
+	case "ItIT":
+		{
+			pickerLocale = "it-IT"
+		}
+	case "PlPL":
+		{
+			pickerLocale = "pl-PL"
+		}
+	case "PtPT":
+		{
+			pickerLocale = "pt-BR"
+		}
+	case "SkSK":
+		{
+			pickerLocale = "sk-SK"
+		}
+	case "RuRU":
+		{
+			pickerLocale = "ru-RU"
+		}
+	default:
+		{
+			pickerLocale = "en-US"
+		}
+	}
+	return pickerLocale
+
+}
+
 func getData(writer http.ResponseWriter, request *http.Request, params httprouter.Params) {
 	var data DataPageInput
 	err := json.NewDecoder(request.Body).Decode(&data)
@@ -131,10 +238,15 @@ func getData(writer http.ResponseWriter, request *http.Request, params httproute
 	}
 	logInfo("MAIN", "Loading data for "+data.Data)
 	logInfo("MAIN", "Number of workplaces selected "+strconv.Itoa(len(data.Workplaces)))
-	logInfo("MAIN", "From "+data.From)
-	logInfo("MAIN", "To "+data.To)
+
+	loc, err := time.LoadLocation(location)
+	if err != nil {
+		logError("MAIN", "Problem loading timezone, setting Europe/Prague")
+		loc, _ = time.LoadLocation("Europe/Prague")
+	}
 	layout := "2006-01-02;15:04:05"
-	dateFrom, err := time.Parse(layout, data.From)
+	dateFrom, err := time.ParseInLocation(layout, data.From, loc)
+	dateFrom = dateFrom.In(time.UTC)
 	if err != nil {
 		logError("MAIN", "Problem parsing date: "+data.From)
 		var responseData DataPageOutput
@@ -144,7 +256,8 @@ func getData(writer http.ResponseWriter, request *http.Request, params httproute
 		logInfo("MAIN", "Processing data ended")
 		return
 	}
-	dateTo, err := time.Parse(layout, data.To)
+	dateTo, err := time.ParseInLocation(layout, data.To, loc)
+	dateTo = dateTo.In(time.UTC)
 	if err != nil {
 		logError("MAIN", "Problem parsing date: "+data.To)
 		var responseData DataPageOutput
@@ -154,8 +267,18 @@ func getData(writer http.ResponseWriter, request *http.Request, params httproute
 		logInfo("MAIN", "Processing data ended")
 		return
 	}
+
+	logInfo("MAIN", "From "+dateFrom.String())
+	logInfo("MAIN", "To "+dateTo.String())
 	email, _, _ := request.BasicAuth()
 	workplaceIds, locale := processDataFromDatabase(data, err)
+
+	userSettingsSync.Lock()
+	settings := cachedUserSettings[email]
+	settings.dataSelection = locale.Name
+	cachedUserSettings[email] = settings
+	userSettingsSync.Unlock()
+
 	switch locale.Name {
 	case "alarms":
 		{
@@ -268,9 +391,9 @@ func getData(writer http.ResponseWriter, request *http.Request, params httproute
 
 			}
 			var orderRecords []database.OrderRecord
-			db.Where(workplaceIds).Where("date_time_start >= ?", dateFrom).Where("date_time_end <= ?", dateTo).Or("date_time_end is null").Find(&orderRecords)
+			db.Where(workplaceIds).Where("date_time_start >= ?", dateFrom).Where("date_time_start <= ?", dateTo).Order("date_time_start desc").Find(&orderRecords)
 			var userRecords []database.UserRecord
-			db.Where(workplaceIds).Where("date_time_start >= ?", dateFrom).Where("date_time_end <= ?", dateTo).Or("date_time_end is null").Find(&userRecords)
+			db.Where(workplaceIds).Where("date_time_start >= ?", dateFrom).Where("date_time_start <= ?", dateTo).Find(&userRecords)
 			var userRecordsByRecordId = map[int]database.UserRecord{}
 			for _, record := range userRecords {
 				userRecordsByRecordId[record.OrderRecordID] = record
@@ -296,50 +419,92 @@ func getData(writer http.ResponseWriter, request *http.Request, params httproute
 			logInfo("MAIN", "Sending "+strconv.Itoa(len(orderRecordsOutput))+" orders")
 			var data TableData
 
-			// TODO: create locales for order table headers
-			// TODO: create locales for data-table-rows-count-title	Show entries:	Title for rows steps box
-			// TODO: create locales for data-table-search-title	Search:	Title for search input
-			// TODO: create locales for data-table-info-title	Showing $1 to $2 of $3 entries	Title for table info block
-			// TODO: create locales for data-pagination-prev-title	Prev	Title pagination prev button
-			// TODO: create locales for data-pagination-next-title	Next	Title pagination next button
-			// TODO: create locales for data-all-records-title	All	Title all records in rows steps block
-			// TODO: create locales for data-inspector-title	Inspector	Title for table inspector window
+			data.DataTableSearchTitle = getLocale(email, "data-table-search-title")
+			data.DataTableInfoTitle = getLocale(email, "data-table-info-title")
+			data.DataTableRowsCountTitle = getLocale(email, "data-table-rows-count-title")
 
-			// TODO: creates headers
-			var headerCell HeaderCell
-			headerCell.HeaderName = "Workplace Name"
-			data.TableHeader = append(data.TableHeader, headerCell)
-			var headerCell2 HeaderCell
-			headerCell2.HeaderName = "b"
-			data.TableHeader = append(data.TableHeader, headerCell2)
+			workplaceName := HeaderCell{HeaderName: getLocale(email, "workplace-name")}
+			data.TableHeader = append(data.TableHeader, workplaceName)
+			orderStart := HeaderCell{HeaderName: getLocale(email, "order-start")}
+			data.TableHeader = append(data.TableHeader, orderStart)
+			orderEnd := HeaderCell{HeaderName: getLocale(email, "order-end")}
+			data.TableHeader = append(data.TableHeader, orderEnd)
+			workplaceModeName := HeaderCell{HeaderName: getLocale(email, "workplacemode-name")}
+			data.TableHeader = append(data.TableHeader, workplaceModeName)
+			workshiftName := HeaderCell{HeaderName: getLocale(email, "workshift-name")}
+			data.TableHeader = append(data.TableHeader, workshiftName)
+			userName := HeaderCell{HeaderName: getLocale(email, "user-name")}
+			data.TableHeader = append(data.TableHeader, userName)
+			orderName := HeaderCell{HeaderName: getLocale(email, "order-name")}
+			data.TableHeader = append(data.TableHeader, orderName)
+			operationName := HeaderCell{HeaderName: getLocale(email, "operation-name")}
+			data.TableHeader = append(data.TableHeader, operationName)
+			cycleName := HeaderCell{HeaderName: getLocale(email, "cycle-name")}
+			data.TableHeader = append(data.TableHeader, cycleName)
+			cavityName := HeaderCell{HeaderName: getLocale(email, "cavity-name")}
+			data.TableHeader = append(data.TableHeader, cavityName)
+			goodPcsName := HeaderCell{HeaderName: getLocale(email, "good-pieces-name")}
+			data.TableHeader = append(data.TableHeader, goodPcsName)
+			badPcsName := HeaderCell{HeaderName: getLocale(email, "bad-pieces-name")}
+			data.TableHeader = append(data.TableHeader, badPcsName)
+			noteName := HeaderCell{HeaderName: getLocale(email, "note-name")}
+			data.TableHeader = append(data.TableHeader, noteName)
 
 			// TODO: loop data and add rows
-			//for _, record := range orderRecords {
-			//var tableRow1 TableRow
-			//var tableRow2 TableRow
-			//
-			//var tableCellA1 TableCell
-			//tableCellA1.CellName = "testA1"
-			//
-			//var tableCellA2 TableCell
-			//tableCellA2.CellName = "testA2"
-			//
-			//var tableCellB1 TableCell
-			//tableCellB1.CellName = "testB1"
-			//
-			//var tableCellB2 TableCell
-			//tableCellB2.CellName = "testB2"
-			//
-			//tableRow1.TableCell = append(tableRow1.TableCell, tableCellA1)
-			//tableRow1.TableCell = append(tableRow1.TableCell, tableCellA2)
-			//
-			//tableRow2.TableCell = append(tableRow2.TableCell, tableCellB1)
-			//tableRow2.TableCell = append(tableRow2.TableCell, tableCellB2)
-			//
-			//data.TableRows = append(data.TableRows, tableRow1)
-			//data.TableRows = append(data.TableRows, tableRow2)
-			//}
+			for _, record := range orderRecords {
+				var tableRow TableRow
 
+				workplaceNameCell := TableCell{CellName: cachedWorkplacesById[uint(record.WorkplaceID)].Name}
+				tableRow.TableCell = append(tableRow.TableCell, workplaceNameCell)
+
+				orderStart := TableCell{CellName: record.DateTimeStart.Format("2006-01-02 15:04:05")}
+				tableRow.TableCell = append(tableRow.TableCell, orderStart)
+
+				if record.DateTimeEnd.Time.IsZero() {
+					orderEnd := TableCell{CellName: time.Now().Format("2006-01-02 15:04:05") + " +"}
+					tableRow.TableCell = append(tableRow.TableCell, orderEnd)
+				} else {
+					orderEnd := TableCell{CellName: record.DateTimeEnd.Time.Format("2006-01-02 15:04:05")}
+					tableRow.TableCell = append(tableRow.TableCell, orderEnd)
+				}
+				workplaceModeNameCell := TableCell{CellName: cachedWorkplaceModesById[uint(record.WorkplaceModeID)].Name}
+				tableRow.TableCell = append(tableRow.TableCell, workplaceModeNameCell)
+
+				workshiftName := TableCell{CellName: cachedWorkshiftsById[uint(record.WorkshiftID)].Name}
+				tableRow.TableCell = append(tableRow.TableCell, workshiftName)
+
+				actualUserId := userRecordsByRecordId[int(record.ID)].UserID
+				userName := TableCell{CellName: cachedUsersById[uint(actualUserId)].FirstName + " " + cachedUsersById[uint(actualUserId)].SecondName}
+				tableRow.TableCell = append(tableRow.TableCell, userName)
+
+				orderName := TableCell{CellName: cachedOrdersById[uint(record.OrderID)].Name}
+				tableRow.TableCell = append(tableRow.TableCell, orderName)
+
+				operationName := TableCell{CellName: cachedOperationsById[uint(record.OperationID)].Name}
+				tableRow.TableCell = append(tableRow.TableCell, operationName)
+
+				averageCycleAsString := strconv.FormatFloat(float64(record.AverageCycle), 'f', 2, 64)
+				averageCycle := TableCell{CellName: averageCycleAsString + "s"}
+				tableRow.TableCell = append(tableRow.TableCell, averageCycle)
+
+				cavityAsString := strconv.Itoa(record.Cavity)
+				cavity := TableCell{CellName: cavityAsString}
+				tableRow.TableCell = append(tableRow.TableCell, cavity)
+
+				okAsString := strconv.Itoa(record.CountOk)
+				ok := TableCell{CellName: okAsString}
+				tableRow.TableCell = append(tableRow.TableCell, ok)
+
+				nokAsString := strconv.Itoa(record.CountNok)
+				nok := TableCell{CellName: nokAsString}
+				tableRow.TableCell = append(tableRow.TableCell, nok)
+
+				note := TableCell{CellName: record.Note}
+				tableRow.TableCell = append(tableRow.TableCell, note)
+
+				data.TableRows = append(data.TableRows, tableRow)
+
+			}
 			tmpl := template.Must(template.ParseFiles("./html/table.html"))
 			_ = tmpl.Execute(writer, data)
 			logInfo("MAIN", "Home page sent for "+email)