+++ js/data.js
@@ -1,11 +1,11 @@
+var locales = Metro.utils.getLocales();
+console.log(locales);
+
 let dt = new Date();
 dt = addDate(dt, -1, 'months');
 const datePicker = document.getElementById("fromdate")
 datePicker.dataset.value = dt.format("%Y-%m-%d")
-
 const dataOkButton = document.getElementById("data-ok-button")
-
-
 
 dataOkButton.addEventListener("click", (event) => {
     var locales = Metro.utils.getLocales();
@@ -27,12 +27,28 @@ dataOkButton.addEventListener("click", (event) => {
         body: JSON.stringify(data)
     }).then((response) => {
         response.text().then(function (data) {
+            console.log(data)
             document.getElementById("data-table-container").innerHTML = data
         });
     }).catch((error) => {
         console.log(error)
     });
 })
+
+
+
+// const exportButton = document.getElementById("data-export-button")
+//
+// exportButton.addEventListener("click", (event) => {
+//     console.log("clicked")
+//     var table = $(el).data('data-table-container');
+//     console.log(table)
+//     table.export('CSV', 'all', 'table-export.csv', {
+//         csvDelimiter: "\t",
+//         csvNewLine: "\r\n",
+//         includeHeader: true
+//     });
+// })
 
 
 function addDate(dt, amount, dateType) {