+++ js/locales.js
@@ -1,0 +1,102 @@
+Metro.utils.addLocale({
+    'cs-CZ': {
+        "calendar": {
+            "months": [
+                "Leden", "Únor", "Březen", "Duben", "Květen", "Červen", "Červenec", "Srpen", "Září", "Říjen", "Listopad", "Prosinec",
+                "Led", "Úno", "Bře", "Dub", "Kvě", "Čer", "Čvc", "Srp", "Žář", "Říj", "Lis", "Pro"
+            ],
+            "days": [
+                "Neděle", "Pondělí", "Úterý", "Středa", "Čtvrtek", "Pátek", "Sobota",
+                "Ne", "Po", "Út", "St", "Čt", "Pá", "So",
+                "Ned", "Pon", "Úte", "Stř", "Čtv", "Pát", "Sob"
+            ],
+            "time": {
+                "days": "DNŮ",
+                "hours": "HOD",
+                "minutes": "MIN",
+                "seconds": "SEK"
+            }
+        },
+        "buttons": {
+            "ok": "OK",
+            "cancel": "Zrušit",
+            "done": "Hotovo",
+            "today": "Dnes",
+            "now": "Nyní",
+            "clear": "Vyčistit",
+            "help": "Pomoc",
+            "yes": "Ano",
+            "no": "Ne",
+            "random": "Náhodné"
+        }
+    }
+});
+
+Metro.utils.addLocale({
+    'sk-SK': {
+        "calendar": {
+            "months": [
+                "Január", "Február", "Marec", "Apríl", "Máj", "Jún", "Júl", "August", "September", "Október", "November", "December",
+                "Jan", "Feb", "Bar", "Apr", "Maj", "Jun", "Jul", "Aug", "Sep", "Okt", "Nov", "Dec"
+            ],
+            "days": [
+                "Nedeľa", "Pondelok", "Utorok", "Streda", "Štvrtok", "Piatok", "Sobota",
+                "Ne", "Po", "Ut", "St", "Št", "Pi", "So",
+                "Ned", "Pon", "Uto", "Str", "Štv", "Pia", "Sob"
+            ],
+            "time": {
+                "days": "DNÍ",
+                "hours": "HOD",
+                "minutes": "MIN",
+                "seconds": "SEK"
+            }
+        },
+        "buttons": {
+            "ok": "OK",
+            "cancel": "Zrušit",
+            "done": "Hotovo",
+            "today": "Dnes",
+            "now": "Nyní",
+            "clear": "Vyčistit",
+            "help": "Pomoc",
+            "yes": "Ano",
+            "no": "Ne",
+            "random": "Náhodné"
+        }
+    }
+});
+
+
+Metro.utils.addLocale({
+    'pl-PL': {
+        "calendar": {
+            "months": [
+                "Styczeń", "Luty", "Marzec", "Kwiecień", "Maj", "Czerwiec", "Lipiec", "Sierpień", "Wrzesień", "Październik", "Listopad", "Grudzień",
+                "Sty", "Lut", "Mar", "Kwi", "Maj", "Cze", "Lip", "Sie", "Wrz", "Paź", "Lis", "Gru"
+            ],
+            "days": [
+                "Niedziela", "Poniedziałek", "Wtorek", "Środa", "Czwartek", "Piątek", "Sobota",
+                "Ni", "Po", "Wt", "Śr", "Cz", "Pi", "So",
+                "Nie", "Pon", "Wto", "Śro", "Czw", "Pią", "Sob"
+            ],
+            "time": {
+                "days": "DNI",
+                "hours": "GOD",
+                "minutes": "MIN",
+                "seconds": "SEK"
+            }
+        },
+        "buttons": {
+            "ok": "OK",
+            "cancel": "Zrušit",
+            "done": "Hotovo",
+            "today": "Dnes",
+            "now": "Nyní",
+            "clear": "Vyčistit",
+            "help": "Pomoc",
+            "yes": "Ano",
+            "no": "Ne",
+            "random": "Náhodné"
+        }
+    }
+});